refactor to use wrapper function to wrap all the codes in cast_string_to_function.h
