+++ dataset/load-from-test/fixed_list/fixed_list_correct.csv
@@ -1,6 +1,6 @@
-"[23123.4342133123]"
-"[-4321312.42343255]"
-"[43213123454.34324]"
-"[-1.344455]"
-"[-4212.4234]"
-"[-2]"
+"[23123.4342133123]","T","1 month 2 days"
+"[-4321312.42343255]",true,1 year 3 days
+"[43213123454.34324]",false,12 years 3 months 30 days
+"[-1.344455]",F,3 days
+"[-4212.4234]",0,null
+"[-2]",1,10 months
