+++ dataset/load-from-test/fixed_list/fixed_list_int64.csv
@@ -1,3 +1,3 @@
-"[4234234, 4234535, -3242342]"
-"[4234234, 4234535, 0]"
-"[4234234, 0, -3242342]"
+"[4234234, 4234535, -3242342]","2022-06-06","2022-11-22 15:12:22"
+"[4234234, 4234535, 0]","2013-02-21","1970-01-01 00:00:00.004666-10"
+"[4234234, 0, -3242342]","2019-12-30","1970-01-01 10:11:00.004666+10:11"
