+++ src/function/CMakeLists.txt
@@ -7,7 +7,8 @@ add_library(kuzu_function
         built_in_aggregate_functions.cpp
         built_in_vector_functions.cpp
         built_in_table_functions.cpp
-        cast_utils.cpp
+        cast_string_non_nested_functions.cpp
+        cast_string_to_functions.cpp
         comparison_functions.cpp
         find_function.cpp
         scalar_macro_function.cpp
