+++ src/include/function/cast/functions/cast_string_non_nested_functions.h
@@ -3,14 +3,13 @@
 #include <cassert>
 
 #include "common/exception/conversion.h"
-#include "common/exception/overflow.h"
 #include "common/string_utils.h"
 #include "common/type_utils.h"
 #include "common/types/int128_t.h"
 #include "common/types/ku_string.h"
 #include "common/vector/value_vector.h"
 #include "fast_float.h"
-#include "numeric_limits.h"
+#include "numeric_cast.h"
 
 namespace kuzu {
 namespace function {
@@ -39,8 +38,8 @@ struct IntegerCastOperation {
         }
         return true;
     }
-    // TODO: handle decimals
 
+    // TODO(Kebing): handle decimals
     template<class T, bool NEGATIVE>
     static bool finalize(T& /*state*/) {
         return true;
@@ -52,7 +51,7 @@ bool tryCastToBool(const char* input, uint64_t len, bool& result);
 void castStringToBool(const char* input, uint64_t len, bool& result);
 
 // cast to numerical values
-// TODO: support exponent + decimal
+// TODO(Kebing): support exponent + decimal
 template<typename T, bool NEGATIVE, bool ALLOW_EXPONENT = false, class OP>
 static bool integerCastLoop(const char* input, uint64_t len, T& result) {
     int64_t start_pos = 0;
@@ -62,7 +61,6 @@ static bool integerCastLoop(const char* input, uint64_t len, T& result) {
     int64_t pos = start_pos;
     while (pos < len) {
         if (!common::StringUtils::CharacterIsDigit(input[pos])) {
-            // TODO: exponent and decimals
             return false;
         }
         uint8_t digit = input[pos++] - '0';
@@ -174,7 +172,7 @@ static bool trySimpleInt128Cast(const char* input, uint64_t len, common::int128_
 
 static void simpleInt128Cast(const char* input, uint64_t len, common::int128_t& result) {
     if (!trySimpleInt128Cast(input, len, result)) {
-        throw common::OverflowException(
+        throw common::ConversionException(
             "Cast failed. " + std::string{input, len} + " is not within INT128 range.");
     }
 }
@@ -228,216 +226,6 @@ static void doubleCast(const char* input, uint64_t len, T& result,
             "Cast failed. " + std::string{input, len} + " is not in " +
             common::LogicalTypeUtils::dataTypeToString(type) + " range.");
     }
-}
-
-template<typename T>
-static inline T castStringToNum(const char* input, uint64_t len,
-    const common::LogicalType& type = common::LogicalType{common::LogicalTypeID::ANY}) {
-    T result;
-    simpleIntegerCast(input, len, result, type);
-    return result;
-}
-
-template<>
-inline common::int128_t castStringToNum(const char* input, uint64_t len,
-    const common::LogicalType& /*type*/) { // NOLINT(misc-unused-parameters): False positive
-    common::int128_t result{};
-    simpleInt128Cast(input, len, result);
-    return result;
-}
-
-template<>
-inline uint64_t castStringToNum(const char* input, uint64_t len, const common::LogicalType& type) {
-    uint64_t result;
-    simpleIntegerCast<uint64_t, false>(input, len, result, type);
-    return result;
-}
-
-template<>
-uint32_t castStringToNum(const char* input, uint64_t len, const common::LogicalType& type) {
-    uint32_t result;
-    simpleIntegerCast<uint32_t, false>(input, len, result, type);
-    return result;
-}
-
-template<>
-uint16_t castStringToNum(const char* input, uint64_t len, const common::LogicalType& type) {
-    uint16_t result;
-    simpleIntegerCast<uint16_t, false>(input, len, result, type);
-    return result;
-}
-
-template<>
-uint8_t castStringToNum(const char* input, uint64_t len, const common::LogicalType& type) {
-    uint8_t result;
-    simpleIntegerCast<uint8_t, false>(input, len, result, type);
-    return result;
-}
-
-template<>
-inline double_t castStringToNum(const char* input, uint64_t len, const common::LogicalType& type) {
-    double_t result;
-    doubleCast<double_t>(input, len, result, type);
-    return result;
-}
-
-template<>
-inline float_t castStringToNum(const char* input, uint64_t len, const common::LogicalType& type) {
-    float_t result;
-    doubleCast<float_t>(input, len, result, type);
-    return result;
-}
-
-template<class SRC, class DST>
-static bool tryCastWithOverflowCheck(SRC value, DST& result) {
-    if (NumericLimits<SRC>::isSigned() != NumericLimits<DST>::isSigned()) {
-        if (NumericLimits<SRC>::isSigned()) {
-            if (NumericLimits<SRC>::digits() > NumericLimits<DST>::digits()) {
-                if (value < 0 || value > (SRC)NumericLimits<DST>::maximum()) {
-                    return false;
-                }
-            } else {
-                if (value < 0) {
-                    return false;
-                }
-            }
-            result = (DST)value;
-            return true;
-        } else {
-            // unsigned to signed conversion
-            if (NumericLimits<SRC>::digits() >= NumericLimits<DST>::digits()) {
-                if (value <= (SRC)NumericLimits<DST>::maximum()) {
-                    result = (DST)value;
-                    return true;
-                }
-                return false;
-            } else {
-                result = (DST)value;
-                return true;
-            }
-        }
-    } else {
-        // same sign conversion
-        if (NumericLimits<DST>::digits() >= NumericLimits<SRC>::digits()) {
-            result = (DST)value;
-            return true;
-        } else {
-            if (value < SRC(NumericLimits<DST>::minimum()) ||
-                value > SRC(NumericLimits<DST>::maximum())) {
-                return false;
-            }
-            result = (DST)value;
-            return true;
-        }
-    }
-}
-
-template<class SRC, class T>
-bool tryCastWithOverflowCheckFloat(SRC value, T& result, SRC min, SRC max) {
-    if (!(value >= min && value < max)) {
-        return false;
-    }
-    // PG FLOAT => INT casts use statistical rounding.
-    result = std::nearbyint(value);
-    return true;
-}
-
-template<>
-bool tryCastWithOverflowCheck(float value, int8_t& result) {
-    return tryCastWithOverflowCheckFloat<float, int8_t>(value, result, -128.0f, 128.0f);
-}
-
-template<>
-bool tryCastWithOverflowCheck(float value, int16_t& result) {
-    return tryCastWithOverflowCheckFloat<float, int16_t>(value, result, -32768.0f, 32768.0f);
-}
-
-template<>
-bool tryCastWithOverflowCheck(float value, int32_t& result) {
-    return tryCastWithOverflowCheckFloat<float, int32_t>(
-        value, result, -2147483648.0f, 2147483648.0f);
-}
-
-template<>
-bool tryCastWithOverflowCheck(float value, int64_t& result) {
-    return tryCastWithOverflowCheckFloat<float, int64_t>(
-        value, result, -9223372036854775808.0f, 9223372036854775808.0f);
-}
-
-template<>
-bool tryCastWithOverflowCheck(float value, uint8_t& result) {
-    return tryCastWithOverflowCheckFloat<float, uint8_t>(value, result, 0.0f, 256.0f);
-}
-
-template<>
-bool tryCastWithOverflowCheck(float value, uint16_t& result) {
-    return tryCastWithOverflowCheckFloat<float, uint16_t>(value, result, 0.0f, 65536.0f);
-}
-
-template<>
-bool tryCastWithOverflowCheck(float value, uint32_t& result) {
-    return tryCastWithOverflowCheckFloat<float, uint32_t>(value, result, 0.0f, 4294967296.0f);
-}
-
-template<>
-bool tryCastWithOverflowCheck(float value, uint64_t& result) {
-    return tryCastWithOverflowCheckFloat<float, uint64_t>(
-        value, result, 0.0f, 18446744073709551616.0f);
-}
-
-template<>
-bool tryCastWithOverflowCheck(double value, int8_t& result) {
-    return tryCastWithOverflowCheckFloat<double, int8_t>(value, result, -128.0, 128.0);
-}
-
-template<>
-bool tryCastWithOverflowCheck(double value, int16_t& result) {
-    return tryCastWithOverflowCheckFloat<double, int16_t>(value, result, -32768.0, 32768.0);
-}
-
-template<>
-bool tryCastWithOverflowCheck(double value, int32_t& result) {
-    return tryCastWithOverflowCheckFloat<double, int32_t>(
-        value, result, -2147483648.0, 2147483648.0);
-}
-
-template<>
-bool tryCastWithOverflowCheck(double value, int64_t& result) {
-    return tryCastWithOverflowCheckFloat<double, int64_t>(
-        value, result, -9223372036854775808.0, 9223372036854775808.0);
-}
-
-template<>
-bool tryCastWithOverflowCheck(double value, uint8_t& result) {
-    return tryCastWithOverflowCheckFloat<double, uint8_t>(value, result, 0.0, 256.0);
-}
-
-template<>
-bool tryCastWithOverflowCheck(double value, uint16_t& result) {
-    return tryCastWithOverflowCheckFloat<double, uint16_t>(value, result, 0.0, 65536.0);
-}
-
-template<>
-bool tryCastWithOverflowCheck(double value, uint32_t& result) {
-    return tryCastWithOverflowCheckFloat<double, uint32_t>(value, result, 0.0, 4294967296.0);
-}
-
-template<>
-bool tryCastWithOverflowCheck(double value, uint64_t& result) {
-    return tryCastWithOverflowCheckFloat<double, uint64_t>(
-        value, result, 0.0, 18446744073709551615.0);
-}
-
-template<>
-bool tryCastWithOverflowCheck(float input, double& result) {
-    result = double(input);
-    return true;
-}
-
-template<>
-bool tryCastWithOverflowCheck(double input, float& result) {
-    result = float(input);
-    return true;
 }
 
 } // namespace function
