+++ src/processor/operator/persistent/copy_to_csv.cpp
@@ -28,7 +28,7 @@ void CopyToCSVLocalState::init(CopyToInfo* info, MemoryManager* mm, ResultSet* r
     castFuncs.resize(info->dataPoses.size());
     for (auto i = 0u; i < info->dataPoses.size(); i++) {
         auto vectorToCast = resultSet->getValueVector(info->dataPoses[i]);
-        function::CastToStringVectorFunction::getUnaryCastExecFunction(
+        function::CastToStringVectorFunction::getUnaryCastToStringExecFunction(
             vectorToCast->dataType.getLogicalTypeID(), castFuncs[i]);
         vectorsToCast.push_back(std::move(vectorToCast));
         auto castVector = std::make_unique<ValueVector>(LogicalTypeID::STRING, mm);
